+++ src/components/register/selectBox/errorMessage/ErrorMessage.Sytle.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { Colors, FontStyles } from '../../../styles';
+import { Colors, FontStyles } from '../../../../styles';
 
 export const ErrorMessage = styled.div`
   display: flex;
